requirements.txt
python-nmap
shodan
requests
